# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is API Documentation",
    "version": "1.0.0",
    "title": "ToDo App",
    "contact": {
      "email": "abduroza@students.unnes.ac.id"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/task/update/{id}": {
      "put": {
        "tags": [
          "Task"
        ],
        "summary": "Update Todo",
        "description": "",
        "operationId": "updateTodo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Task will be update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TodoUpdate"
            }
          },
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "description": "task will update"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo updated"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "TodoUpdate": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jogging"
        },
        "date-time": {
          "type": "string",
          "example": "2019-08-28T10:30:00.000Z"
        },
        "note": {
          "type": "string",
          "example": "biar gembira"
        },
        "priority": {
          "type": "string",
          "description": "task priority",
          "enum": [
            "none",
            "low",
            "medium",
            "high"
          ],
          "default": "none",
          "example": "high"
        },
        "status": {
          "type": "boolean",
          "default": false,
          "example": true
        }
      }
    }
  },
...
}